sounds: Pick ringtone using webos-lib FilePicker
The existing code seemed to be pencilling in a roll-your-own version
with fake content and a "Someone else should write this" idea.
Let's not write it at all.
Slight drawback: FilePicker does not work very well (at the moment?).
Hardly seems to work at all for "audio". That's why I have not attempted
to actually set the choice in the system yet.
Can usually manage to pick a new wallpaper with it so perhaps there is
something funny about "audio" (at the moment?).